Fix grub installation when DEBUG=0.
update-grub was failing in the installation process, when DEBUG=0.
When DEBUG=1, the line 'GRUB_TERMINAL=serial' of the file /etc/default/grub
was preventing the error to occur. (this is very strange)
Anyway, at one point update-grub was trying to retrieve the path of the
loop device where the squashfs filesystem image was mapped. Since we had done
a chroot in the meanwhile, the path returned by the kernel was not valid anymore,
and this was causing update-grub to fail.
In order to avoid that, we apply the chroot before mounting the squashfs image, and
thus update-grub is happy.