Work appropriately with other shells (not bash)
Replaced bash-specific statements.

Handle ctrl-c appropriately on non-bash shells: bash will always call the
EXIT trap, even if the exit was due to a signal, whereas other shells tend
to call EXIT only in case of an error (with set -e) or normal exit, not when
a signal caused the exit.
The SIGINT signal (called when ctrl-c is pressed) is now specifically handled
in order to work around this issue, and have the same behaviour whatever is
the shell.