Fix image size indicated in the final message.
The image size generated contains holes, and thus
the room taken on the filesystem is smaller than the
'apparent' size (as reported by 'ls') of the image
file.
Previously we were printing the size taken on the
filesystem, which is confusing: the size of the device
where the image will be copied should match at least
the 'apparent' size.
Thus we now print the 'apparent' size.