Tune LVM config during startup.
New versions of LVM (seen on stretch) come with a default
configuration where LVM commands will rely on a set of
daemons (lvmetad for metadata, udev, etc.). Since these
daemons are not started when init scripts are run, we
modify this startup configuration and restore the user
provided config when the startup scripts are done.