Fix error detection in grub install function.
Installation is performed as the 1st command of a pipeline,
and the return code was not caught properly.

The most simple fix would be to activate the pipefail option,
but some shells do not support it (including dash, the default
in debian).

The new code replaces the pipeline with a command substitution.

Thanks to Andreas Unterkircher for reporting the issue.