Enhance --config-kernel-bootargs: add or remove bootargs
--config-kernel-bootargs option may now be used to remove
a kernel bootarg in grub configuration, by prepending a minus sign.

For example --config-kernel-bootargs '+console=ttyS0 -quiet'
will add 'console=ttyS0' but remove 'quiet'.

The plus sign is optional. If no sign is present, the bootarg is added
(i.e. '+console=ttyS0' or just 'console=ttyS0' have the same effect).
Because of this this change is compatible with previous versions.