failsafe: allow failsafe commands after pipes
Previously, this kind of code would not work properly:

start_failsafe_mode
[...]
<cmd> | {
    [...]
    failsafe <cmd>
}

This is because the list of failsafe commands was maintained
in an environment variable. Thus, since 'failsafe <cmd>' occurs
after the pipe, it is run in a subprocess, and modifications of
this environment variable would be lost after the pipe ends.

This environment variable has been replaced by a file, thus
allowing this kind of construct.