disk layouts: handle fixed size for partitions.
Except for the last partition, the generated disk image will have
to include the room needed for this partition, so this option is
not recommended.
Instead, one should favor:
* fixed size on an lvm volume
* min size on this partition (size='auto')
However, sometimes one may want to reserve a little more space
than needed initially, for instance on a /boot partition that will
receive kernel image updates.