Improve /etc/resolv.conf handling.
debootstick needs a valid resolv.conf file in the chroot to be able
to download packages.

Previously, we were bind mounting the host resolv.conf file during
the draft image generation step. This can cause problems if debootstick
installs a package involved with DNS management (such as systemd)
and that package tries to replace the file with a symlink as part
of its installation. Because of the bind-mount, this was failing.

Copying host configuration can cause problems too. For instance,
the host OS might rely on a resolver service listening on localhost,
thus resolv.conf will contain a line such as 'nameserver 127.0.0.1',
whereas the guest OS will not provide such a service when booted.

This commit checks if the resolv.conf file is missing in the chroot,
or if it is a file but it is missing a 'nameserver' directive.
In this case, the file is overridden with a configuration targetting
publicly available DNS servers (cloudflare, google, opendns).

Other cases (file is a symlink for instance) are considered OK and
left untouched.

Related to #28.