New option "--sector-size"
This allows to generate an image to be flashed on a disk
with non-default logical sector size. Default is 512 bytes.

Important note: when using the "--system-type installer" option,
whichever will be the target disk sector size after migrating,
you should probably use the default 512-bytes sector size.
The reason for this is that you will flash the image on a USB
device first, and this device probably expects a 512-bytes sector
size.

This option should only be used if you want to flash the image
directly on a 4Kn disk for instance.