Secure-Boot suppport.
Previously, the "primary" boot mode was BIOS, thanks to the
installation of grub-pc package in the embedded OS.
UEFI was made available by a standalone grub image built
separately and written in the EFI partition.
This standalone grub image was retrieving /boot/grub/grub.conf
from the root filesystem partition and executing it, allowing
the UEFI boot path to be updated too when `update-grub` is called
in the guest OS.

This commits installs both boot modes in the guest OS. This is
achieved by installing grub-pc, shim-signed and
grub-efi-[amd64|ia32]-signed packages on the embedded OS. Those
packages do not conflict (grub-efi would).
Then two grub-install commands are run to manage both modes (BIOS
and UEFI). Thanks to shim-signed and grub-efi-[amd64|ia32]-signed,
UEFI with secure-boot is now working too.
With this setup there is no longer a need to build a standalone
UEFI image.

Fixes #33.