Fix /etc/resolv.conf handling
Commit 68fe9de976c1d does not handle properly ubuntu focal case where the
provided chroot already has a symlink /etc/resolv.conf targeting
../run/systemd/resolve/stub-resolv.conf. When working in the chroot,
systemd is not running, so this symlink is broken.

This commit makes a distinction between the final configuration that
should be found when the system is booted (keep this symlink untouched
in this case) and the fact we need a valid DNS configuration when
downloading packages from the chroot. In this case, we move the provided
/etc/resolv.conf file, replace it with an appropriate configuration for
the chroot, and restore it after the packages are installed.