debootstick is now an architecture-independent Debian package
Since we introduced SecureBoot support (commit 0ac6b2039061d1008890f4ee352906366395aa17),
we use grub-install in the target to install grub both in BIOS
and UEFI modes.
Previously, we were generating a standalone UEFI binary. This
operation required proper grub tooling on the host, and
because of that debootstick had dependencies on packages
grub-efi-amd64-bin and grub-efi-ia32-bin. Since these
packages are architecture-dependent, debootstick was
architecture-dependent too: it was built has an amd64 or
i386 package.
This commit updates debian/control to remove these
obsolete dependencies and set the package architecture
to "all".